mysql database functional
