+++ backend/config/db.config.js
@@ -1,8 +1,8 @@
 module.exports = {
 	HOST: "localhost",
-	USER: "root",
-	PASSWORD: "123456",
-	DB: "testdb",
+	USER: "admin",
+	PASSWORD: "admin",
+	DB: "p7",
 	dialect: "mysql",
 	pool: {
 		max: 5,
