+++ backend/models/user.js
@@ -1,27 +1,6 @@
-const dbConfig = require("../config/db.config.js");
-
-const Sequelize = require("sequelize");
-const sequelize = new Sequelize(dbConfig.DB, dbConfig.USER, dbConfig.PASSWORD, {
-	host: dbConfig.HOST,
-	dialect: dbConfig.dialect,
-	operatorsAliases: false,
-
-	pool: {
-		max: dbConfig.pool.max,
-		min: dbConfig.pool.min,
-		acquire: dbConfig.pool.acquire,
-		idle: dbConfig.pool.idle
-	}
-});
-
-const db = {};
-
-db.Sequelize = Sequelize;
-db.sequelize = sequelize;
-
-User.init({
-	//attributes
-	pseudo: {
+module.exports = (sequelize, Sequelize) => {
+	const User = sequelize.define("user", {
+			pseudo: {
 		type: Sequelize.STRING,
 		allowNull: false,
 		unique: true
@@ -30,8 +9,8 @@ User.init({
 		type: Sequelize.STRING,
 		allowNull: false,
 		unique: false
-	},
-}, {
-	sequelize,
-	modelName: 'user'
+	}
 });
+
+return User;
+};
